+++ app-flavor-liveserver/src/main/java/com/handstandsam/shoppingapp/ServerDimension.kt
@@ -1,0 +1,8 @@
+package com.handstandsam.shoppingapp
+
+import android.app.Application
+import com.handstandsam.shoppingapp.di.NetworkGraph
+
+fun Application.serverDimensionNetworkGraph(): NetworkGraph {
+    return LiveNetworkGraph(applicationContext)
+}
