+++ app-flavor-mockserver/src/main/java/com/handstandsam/shoppingapp/network/MockNetworkManager.kt
