--- app/src/mock/java/com/handstandsam/shoppingapp/ServerDimension.kt
@@ -1,25 +1,0 @@
-package com.handstandsam.shoppingapp
-
-import com.handstandsam.shoppingapp.di.BaseNetworkGraph
-import com.handstandsam.shoppingapp.di.NetworkGraph
-import com.handstandsam.shoppingapp.mockdata.ProduceMockAccount
-import com.handstandsam.shoppingapp.network.MockNetworkManager
-
-
-fun MyAbstractApplication.serverDimensionNetworkGraph(): NetworkGraph {
-
-    val networkConfig = NetworkConfigs.LOCALHOST
-
-    val mockAccount = ProduceMockAccount()
-
-    MockNetworkManager(
-        appContext = applicationContext,
-        networkConfig = networkConfig,
-        mockAccount = mockAccount
-    )
-
-    return BaseNetworkGraph(
-        networkConfig = networkConfig,
-        appContext = applicationContext
-    )
-}
