+++ debug/build.gradle
@@ -22,12 +22,17 @@ android {
 }
 
 dependencies {
-    implementation project(":mock-data")
+    api project(":mock-data")
+    implementation project(":networking")
     implementation project(":models")
+
     implementation Libs.timber
     implementation Libs.okhttp
     implementation Libs.apache_commons
     implementation Libs.kotlin_std_lib
+
+    implementation Libs.moshi
+    implementation Libs.moshiKotlin
 
     api Libs.stetho
     api Libs.stetho_okhttp3
