+++ networking/src/main/java/com/handstandsam/shoppingapp/repository/ItemRepo.kt
@@ -1,0 +1,9 @@
+package com.handstandsam.shoppingapp.repository
+
+import com.handstandsam.shoppingapp.models.Item
+import io.reactivex.Single
+
+interface ItemRepo {
+
+    fun getItemsForCategory(categoryLabel: String): Single<List<Item>>
+}
